+++ CHANGELOG.md
@@ -13,9 +13,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 ## v3.XX.X (2023-XX-XX)
 
-#### :rocket: New Feature
-
-* `core/kv-storage` - added a new engine that builds storage based on a string
+* Added an engine that allows storing data in a single string `core/kv-storage`
 
 ## v3.95.4 (2023-05-25)